Moving files around to make it easier to read.
